+++ packages/flutter/lib/src/editor/mawy_editor.dart
@@ -1108,6 +1108,7 @@ class _MawyEditorState extends State<MawyEditor> {
           canRedo: showSource && !widget.readOnly && _history.value.canRedo,
           onTravel: _travel,
           editable: showSource && !widget.readOnly,
+          inTable: showSource && tableRangeAt(_value, _state.start) != null,
           onInsertTable: _insertTableSized,
           tableAvailable: _tableAvailable,
           headingLevels: _headingLevels,
@@ -1457,6 +1458,7 @@ class _Toolbar extends StatefulWidget {
     required this.canRedo,
     required this.onTravel,
     required this.editable,
+    required this.inTable,
     required this.onInsertTable,
     required this.tableAvailable,
     required this.headingLevels,
@@ -1485,6 +1487,10 @@ class _Toolbar extends StatefulWidget {
   /// Whether there is a source showing that can be changed, which a menu of
   /// things to change in it needs before it is worth opening.
   final bool editable;
+
+  /// Whether the caret is in a table, where the commands that make a block have
+  /// nothing to make. See [blockCommand].
+  final bool inTable;
   final void Function(int columns, int rows) onInsertTable;
   final bool Function(MawyTableCommand) tableAvailable;
   final List<int> headingLevels;
@@ -1770,6 +1776,7 @@ class _ToolbarState extends State<_Toolbar> {
             label: widget.strings.heading,
             tokens: widget.tokens,
             focusNode: next(),
+            enabled: !widget.inTable,
             builder: (VoidCallback close) => MawyToolbarChoice<int>(
               tokens: widget.tokens,
               value: widget.headingLevels.firstWhere(
@@ -1809,6 +1816,7 @@ class _ToolbarState extends State<_Toolbar> {
           tokens: widget.tokens,
           focusNode: next(),
           pressed: commandActive(control.command, widget.state),
+          enabled: !(widget.inTable && blockCommand(control.command)),
           onPressed: () => widget.onCommand?.call(control.command),
         ),
       );